Add alpine own federated repo file
Add alpine specific schema, struct, and unpacking func to ensure specific attributes are handled correctly

Add import test case to verify

Move shared funcs to federated.go
