+++ pkg/artifactory/provider/provider.go
@@ -31,6 +31,7 @@ var productId = "terraform-provider-artifactory/" + Version
 func Provider() *schema.Provider {
 	resourceMap := map[string]*schema.Resource{
 		"artifactory_keypair":                             security.ResourceArtifactoryKeyPair(),
+		"artifactory_federated_alpine_repository":         federated.ResourceArtifactoryFederatedAlpineRepository(),
 		"artifactory_federated_maven_repository":          federated.ResourceArtifactoryFederatedJavaRepository("maven", false),
 		"artifactory_local_nuget_repository":              local.ResourceArtifactoryLocalNugetRepository(),
 		"artifactory_local_maven_repository":              local.ResourceArtifactoryLocalJavaRepository("maven", false),
