+++ pkg/artifactory/resource/repository/federated/resource_artifactory_federated_alpine_repository.go
@@ -1,0 +1,58 @@
+package federated
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/artifactory/resource/repository"
+	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/artifactory/resource/repository/local"
+	"github.com/jfrog/terraform-provider-shared/packer"
+	"github.com/jfrog/terraform-provider-shared/predicate"
+	"github.com/jfrog/terraform-provider-shared/util"
+)
+
+func ResourceArtifactoryFederatedAlpineRepository() *schema.Resource {
+	packageType := "alpine"
+
+	type AlpineFederatedRepositoryParams struct {
+		local.AlpineLocalRepoParams
+		Members []Member `hcl:"member" json:"members"`
+	}
+
+	alpineFederatedSchema := util.MergeMaps(
+		local.AlpineLocalSchema,
+		memberSchema,
+		repository.RepoLayoutRefSchema(rclass, packageType),
+	)
+
+	var unpackFederatedAlpineRepository = func(data *schema.ResourceData) (interface{}, string, error) {
+		repo := AlpineFederatedRepositoryParams{
+			AlpineLocalRepoParams: local.UnpackLocalAlpineRepository(data, rclass),
+			Members:               unpackMembers(data),
+		}
+		return repo, repo.Id(), nil
+	}
+
+	var packAlpineMembers = func(repo interface{}, d *schema.ResourceData) error {
+		members := repo.(*AlpineFederatedRepositoryParams).Members
+		return packMembers(members, d)
+	}
+
+	pkr := packer.Compose(
+		packer.Universal(
+			predicate.Ignore("class", "rclass", "member", "terraform_type"),
+		),
+		packAlpineMembers,
+	)
+
+	constructor := func() (interface{}, error) {
+		return &AlpineFederatedRepositoryParams{
+			AlpineLocalRepoParams: local.AlpineLocalRepoParams{
+				RepositoryBaseParams: local.RepositoryBaseParams{
+					PackageType: packageType,
+					Rclass:      rclass,
+				},
+			},
+		}, nil
+	}
+
+	return repository.MkResourceSchema(alpineFederatedSchema, pkr, unpackFederatedAlpineRepository, constructor)
+}
